requirements.txt
telethon==1.36.0
google-api-python-client==2.94.0
google-auth==2.23.0
google-auth-oauthlib==1.2.0
requests==2.31.0
python-dotenv==1.0.0
aiofiles==23.1.0
tqdm==4.66.1
sqlalchemy==2.0.22
alembic==1.11.1
pydantic==2.5.1
aiohttp==3.9.0
asyncio
